op-fetch-appointments-by-dog-id - move appointment status into bark/enums
