+++ src/lib/bark/enums/appointment-status.ts
@@ -1,6 +1,5 @@
 import { z } from "zod";
 
-// WIP: Move this into bark/enums
 export const APPOINTMENT_STATUS = {
   PENDING: "PENDING",
   REPORTED: "REPORTED",
